feat(edgeless): extract public change text menu
