+++ packages/blocks/src/page-block/edgeless/components/text/types.ts
@@ -1,6 +1,11 @@
+import type { ShapeElement, TextElement } from '@blocksuite/phasor/index.js';
+
 import type { EdgelessShapeTextEditor } from './edgeless-shape-text-editor.js';
 import type { EdgelessTextEditor } from './edgeless-text-editor.js';
 
 export type EdgelessCanvasTextEditor =
   | EdgelessShapeTextEditor
   | EdgelessTextEditor;
+
+export type EdgelessCanvasTextElement = ShapeElement | TextElement;
+export type EdgelessCanvasTextElementType = 'shape' | 'text';
